feat: add Step 1 wizard (Configure Resource) to Deploy tab
- Extract the existing deploy view into ResourceEnvironmentsList so the
  /environments tab can host sub-routes
- Make ResourceEnvironments a thin Routes wrapper with two paths: / for
  the deploy view, /parameters-config for the new wizard
- New ResourceParametersConfigPage renders an RJSF form against the
  (Cluster)ResourceType schema. Save & Close PUTs the Resource via the
  existing platform-resource update path; the controller auto-cuts a
  new ResourceRelease from the spec hash change
- Wire the Set up panel's Configure & Deploy button to navigate into the
  new wizard route
- Tests updated to wrap in MemoryRouter and target the extracted list
